requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
distro==1.9.0
et_xmlfile==2.0.0
feedparser==6.0.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
lxml==6.0.2
openai==2.29.0
openpyxl==3.1.5
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-docx==1.2.0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
schedule==1.2.2
sgmllib3k==1.0.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
